*Changed how tests imported files in base MixDown folder *mdLogger.py: Fix typo and lowercase loggername before checking *mdTarget.py: Fix imports, Add ExitOnFailure to Target.Extract() to assist testing *mdTestUtilities.py: Move functions here that will be used in multiple places/files *test_md[Cvs|Git|Hg|Svn].py: Move common code out to mdTestUtilities.py, fix imports *test_mdTarget.py: First of many tests to come for this file

+++ test/test_mdGit.py
@@ -20,24 +20,22 @@
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
 
-import os, sys, unittest, tempfile
+import os, sys, unittest, mdTestUtilities
+
+if not ".." in sys.path:
+    sys.path.append("..")
+import mdGit, utilityFunctions
 
 class test_mdGit(unittest.TestCase):
-    def _createTempGitRepository(self):
-        repoPath = utilityFunctions.includeTrailingPathDelimiter(tempfile.mkdtemp(prefix="mixdown-"))
-        utilityFunctions.executeSubProcess("git init --quiet", repoPath)
-        utilityFunctions.executeSubProcess("touch testFile", repoPath)
-        utilityFunctions.executeSubProcess("git add testFile", repoPath)
-        utilityFunctions.executeSubProcess("git commit -m message --quiet", repoPath)
-        return repoPath
-
     def test_isGitInstalled(self):
         returnValue = mdGit.isGitInstalled()
-        self.assertEqual(returnValue, True, "Git is not installed in your system, all Git tests will fail.")
+        self.assertEqual(returnValue, True, "Git is not installed on your system.  All Git tests will fail.")
 
     def test_isGitRepo(self):
+        if not mdGit.isGitInstalled():
+            self.fail("Git is not installed on your system.  All Git tests will fail.")
         #Create repository and test if is git repo
-        tempRepo = self._createTempGitRepository()
+        tempRepo = mdTestUtilities.createGitRepository()
         try:
             returnValue = mdGit.isGitRepo(tempRepo)
             self.assertEqual(returnValue, True, "mdGit.isGitRepo(" + tempRepo + ") should have returned true.")
@@ -49,18 +47,18 @@ class test_mdGit(unittest.TestCase):
         self.assertEqual(returnValue, False, "mdGit.isGitRepo(" + falsePath + ") should have returned false.")
 
     def test_gitCheckout(self):
-        tempDir = utilityFunctions.includeTrailingPathDelimiter(tempfile.mkdtemp(prefix="mixdown-"))
-        tempRepo = self._createTempGitRepository()
+        if not mdGit.isGitInstalled():
+            self.fail("Git is not installed on your system.  All Git tests will fail.")
+        tempDir = mdTestUtilities.makeTempDir()
+        tempRepo = mdTestUtilities.createGitRepository()
         try:
             mdGit.gitCheckout(tempRepo, tempDir)
-            returnValue = os.path.exists(tempDir + "testFile")
-            self.assertEqual(returnValue, True, "'testFile' did not exist after mdGit.gitCheckout(" + tempRepo + ") was called.")
+            returnValue = os.path.exists(tempDir + mdTestUtilities.testFileName)
+            self.assertEqual(returnValue, True, "'" + mdTestUtilities.testFileName + "' did not exist after mdGit.gitCheckout(" + tempRepo + ") was called.")
         finally:
             utilityFunctions.removeDir(tempDir)
             utilityFunctions.removeDir(tempRepo)
 
 if __name__ == "__main__":
-    sys.path.append("..")
-    import mdGit, utilityFunctions
     unittest.main()
